+++ src/main/kotlin/com/github/madwareru/intellijmdbook/runconfigurations/MdBookBuildConfigurationFactory.kt
@@ -4,15 +4,15 @@ import com.intellij.execution.configurations.ConfigurationFactory
 import com.intellij.execution.configurations.RunConfiguration
 import com.intellij.openapi.project.Project
 
-class MdBookConfigurationFactory(type: MdBookCommandConfigurationType) : ConfigurationFactory(type) {
+class MdBookBuildConfigurationFactory(type: MdBookCommandConfigurationType) : ConfigurationFactory(type) {
 
     override fun getId(): String = ID
 
     override fun createTemplateConfiguration(project: Project): RunConfiguration {
-        return MdBookCommandConfiguration(project, "MdBook", this)
+        return MdBookBuildConfiguration(project, "MdBook Build", this)
     }
 
     companion object {
-        const val ID: String = "MdBook Command"
+        const val ID: String = "MdBook Build"
     }
 }